version 2.0.0 with sync functionality

+++ src/main/kotlin/com/simplereader/sync/SyncEntity.kt
@@ -38,7 +38,7 @@ data class SyncCheckpointEntity(
 data class DeletedRecordsEntity(
     val tableName: String,      // see SyncTables
     val bookId: String,
-    val idx: Long = -1,         // if tableName=bookmark or highlight
+    val idx: Int = -1,         // if tableName=bookmark or highlight
     val sha256: String? = null, // if tableName=book_data, checksum of the file
     val filesize: Long = 0L,    // if tableName=book_data, size of the file
     val deletedAt : Long,       // timestamp of deletion
